Item.fields is now a dict of (name, generator)
The generators are instantiated after all the items
are loaded in the blueprint. That way, if a generator
depends on another item (like a ForeignKey), it can
see it during its instantiation. Moreover, it knows
that the item as not been created yet (this is not
the case in Generator.generate though).